fix some unrecognised annovar funcs being uncategorised
These are compound functional categories like "ncRNA_exonic;splicing", and
we should really be splitting them up, as we can't combinatorially handle
every combination of different categories explicitly. However the logic is a
bit complex, for example, in the case of "ncRNA_exonic;splicing" we have a
a non coding RNA splicing variant. This is nowhere near as important as a
coding splicing variant so we don't want to merely lump it in as a "splicing"
variant. For now, manually dealing with each combination that we observe
coming out of Annovar.